refactor(core): move list rendering to the CLI
Make the engine's list operation return collection data only.

Render list errors, text, JSON, and summaries in the CLI.

Ensure the public listTests() API no longer writes to host stdout.
