+++ packages/core/AGENTS.md
@@ -52,6 +52,7 @@ Contracts between modules or processes — not readable from any single file.
 
 - `src/core/runTests.ts` stays a coarse orchestrator: planner resolve → construct the executors the plan calls for → drive them → finalize. Every run shape — node-only, browser-only, mixed — walks that one assembly and differs only in which executors the plan produced; don't add a shape branch above or below it. Browser-mode detail lives under `src/core/browser/` behind the `BrowserHostModule` load boundary, never inline in the orchestrator.
 - One planner (`src/core/planner.ts`) is the init barrier for both commands: `runTests` and `listTests` each resolve through `createTestPlanner`, inside which both sides' `modifyRstestConfig` hooks fire (the browser's via a files-only discovery boot), so no executor is constructed against a plan still moving. The browser-side `src/core/browser/runPlanner.ts` is constructed and spent inside it — never hold a second planner to keep in step, and never give `listTests` a project-subset or shard decision of its own: every past list-only filter ("have the hooks run yet?") re-derived settlement from proxy signals and broke one config-shape at a time. The shard banner is planner-owned too — resolution only records counts, and the planner announces once after the barrier.
+- Apart from the planner-owned shard banner above, the engine's list collection never writes to process streams; `rstest list` output is rendered only in `src/cli` from `ListCommandCollectionResult`.
 - The cold-start gate is a planner condition, not an orchestrator branch: zero node projects ⇒ the planner skips `prepareRsbuild` and returns no `nodeBuild`, and the orchestrator constructs a node executor exactly when a build came back. The gate is pinned inside the planner (`tests/core/planner.test.ts` spies on `prepareRsbuild`); re-deriving "is this browser-only?" in `runTests` is the shape the second-assembly regression returns in.
 - Watch is one core-owned loop (`src/core/watchSession.ts`): an executor signals `TestExecutor.onInvalidate` with its scope already resolved, and core answers with reset → `runCycle` → `finalizeRunCycle` → next trace buffer → ready banner. Cycles queue, so two never interleave on the shared `stateManager`; a burst folds into the queued cycle only among same-kind triggers, and a fold unions file lists and nothing else — one trigger kind's state (e.g. `u`'s `updateSnapshot`) must never reach files another trigger chose. A trigger that resolves to no work must not signal.
 - Two tracked fold gaps, not stances: the browser executor drops `updateSnapshot` on reruns (the host reads the live flag per page load), and closing that must keep the fold identity honest — the browser side erases trigger kind, so either the origin crosses the seam or `updateSnapshot` joins the fold-compared options. The node trigger still holds the dev-compile hook for its whole cycle (the Watch mode section tracks it) because its stats pull consumes a compile's changes exactly once.
