+++ packages/core/src/cli/commands.ts
@@ -1,15 +1,12 @@
 import cac, { type CAC, type Command } from 'cac';
-import type {
-  ListCommandOptions,
-  RstestCommand,
-  RstestInstance,
-} from '../types';
+import type { RstestCommand, RstestInstance } from '../types';
 import { color, determineAgent, formatError, isTTY, logger } from '../utils';
 import { buildResolvedRunner, isRelatedRun } from '../core/buildRunner';
 import { exitReporters } from '../reporter';
 import type { PackageInstallerConfirm } from '../utils/packageInstaller';
 import { mirrorExitCode } from './exitCode';
 import type { CommonOptions } from './init';
+import { renderListTests, type ListCommandOptions } from './listRenderer';
 import { showRstest } from './prepare';
 
 export type { CommonOptions } from './init';
@@ -736,13 +733,21 @@ export function createCli(): CAC {
           createRstestContext: createRstest,
         });
 
-        await rstest.listTests({
+        const result = await rstest.listTests({
           filesOnly: options.filesOnly,
-          json: options.json,
-          includeSuites: options.includeSuites,
-          printLocation: options.printLocation,
-          summary: options.summary,
         });
+        try {
+          await renderListTests(result, {
+            rootPath: rstest.context.rootPath,
+            filesOnly: options.filesOnly,
+            json: options.json,
+            includeSuites: options.includeSuites,
+            printLocation: options.printLocation,
+            summary: options.summary,
+          });
+        } finally {
+          await result.close();
+        }
       } catch (err) {
         logger.error('Failed to run Rstest list.');
         logger.error(formatError(err));
